+++ packages/terra-core-docs/src/terra-dev-site/doc/scroll/example/ScrollVertical.jsx
@@ -1,8 +1,6 @@
 import React from 'react';
 import Scroll from 'terra-scroll';
 import classNames from 'classnames/bind';
-import Button from 'terra-button';
-import InputField from 'terra-form-input/lib/InputField';
 import styles from './ScrollCommon.module.scss';
 
 const cx = classNames.bind(styles);
@@ -11,8 +9,8 @@ export default () => (
   <div>
     <p>Overflow-Y Container</p>
     <div className={cx('overflow-wrapper')}>
-      <Scroll>
-        <p className={cx('paragraph-x-wrapper')}>
+      <Scroll tabIndex="0">
+        <p className={cx('paragraph-y-wrapper')}>
           Entering patient details is a crucial process in healthcare facilities that involves capturing and documenting relevant information about
           individuals seeking medical care. Patient details need to be regularly updated to ensure accuracy and relevancy. This may involve periodic
           verification of personal information, updating medical history, and making necessary changes to insurance or contact details. Hospitals may
@@ -22,16 +20,6 @@ export default () => (
           information. This usually includes your full name, contact details (such as phone number and email address), and residential address. It is crucial to
           provide accurate contact information as it will be used for communication and verification purposes.
         </p>
-        <div>
-          <InputField
-            inputId="input-field"
-            label="Enter Patient Name"
-            type="text"
-          />
-        </div>
-        <div>
-          <Button text="Submit" />
-        </div>
       </Scroll>
     </div>
   </div>