+++ packages/terra-core-docs/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+
 * Updated
   * Updated `terra-scroll` to add realistic examples.